requirements.txt
django==2.1.6
djangorestframework==3.9.1
Pillow==5.3.0
pkg-resources==0.0.0
pytz==2018.7
